Remove "shared" folder for now
The "shared" folder is mostly interesting for TypeScript. This repository does not contain any actual TypeScript files yet - it merely has JavaScript and ExtendScript content. It _is_ setup for TypeScript, yes, but nothing makes use of this "feature". For now, leave it out to keep things as simple as possible.